+++ src/JWTRetrieval/JWTRetrievalStrategyInterface.php
@@ -7,11 +7,7 @@ use Symfony\Component\HttpFoundation\Request;
 interface JWTRetrievalStrategyInterface
 {
     /**
-     * @param Request $request
-     *
-     * @return string
-     *
      * @throws JWTNotFoundException
      */
-    public function getToken(Request $request);
+    public function getToken(Request $request): string;
 }